requirements.txt
torch>=1.4.0
pretty_midi>=0.2.9
h5py>=2.9.0
matplotlib>=2.2.4
mir_eval>=0.5
pydub>=0.23.1
joblib>=0.13.2
librosa>=0.7.2
figures>=0.3.16
jams>=0.3.4
numpy==1.19.2
pumpp>=0.5.0
scikit_learn>=0.23.2
tqdm